:wrench: chore: configure private routes
