+++ src/routes/auth.routes.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Switch, Route } from 'react-router-dom';
+
+import SignIn from '../pages/SignIn';
+
+const AuthRoutes: React.FC = () => (
+  <Switch>
+    <Route path="/" exact component={SignIn} />
+  </Switch>
+);
+
+export default AuthRoutes;
