+++ src/routes/index.tsx
@@ -1,0 +1,14 @@
+import React, { useContext } from 'react';
+
+import AuthContext from '../Contexts/auth';
+
+import AuthRoutes from './auth.routes';
+import AppRoutes from './app.routes';
+
+const Routes: React.FC = () => {
+  const { signed } = useContext(AuthContext);
+
+  return signed ? <AppRoutes /> : <AuthRoutes />;
+};
+
+export default Routes;
